Add `alwaysStringify` option
This new feature adds support for forcing succesful files to always
stringify, even if they are not written to `streamOut` or the
file-system.  This boolean (default: `false`), can be used if
processed files are handled after a successful  run.

The default behaviour is to only compile if a file is actually output.